Add stable h1 to site editor edit mode
